updated readme and added tsl fix to github tests
