+++ samples/RestFluencing.Sample.GitHub/GitHubGetRequests.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Net;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using RestFluencing;
 using RestFluencing.Assertion;
@@ -8,6 +9,12 @@ namespace RestFluencing.Sample.GitHub
 	[TestClass]
 	public class GitHubGetRequests
 	{
+		[TestInitialize]
+		public void FixTsl()
+		{
+			ServicePointManager.SecurityProtocol = SecurityProtocolType.Tls12;
+		}
+
 		[TestMethod]
 		public void SimpleGetRequest()
 		{
@@ -16,6 +23,23 @@ namespace RestFluencing.Sample.GitHub
 				.Response()
 				.ReturnsStatus(HttpStatusCode.OK)
 				.Assert();
+		}
+
+		[TestMethod]
+		public void GetTheModelToPassItOn()
+		{
+			GitHubUserModel user = null;
+			Rest.GetFromUrl("https://api.github.com/users/defunkt")
+				.WithHeader("User-Agent", "RestFluencing Sample")
+				.Response()
+				.ReturnsModel<GitHubUserModel>(model =>
+				{
+					user = model;
+					return true;
+				}, string.Empty)
+				.Assert();
+
+			Assert.IsNotNull(user);
 		}
 
 	}
